+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add CAISO LMP Heat Map Example Notebook
 - Add a guide on contributing to gridstatus
+- Add Settlement Point Prices for ERCOT
 
 ## v0.16.0 - Dec 15, 2022